POSIX support for syncrhonized websocket
To build on POSIX, I needed to not assume that pthread_mutext_t was a
pointer.  I simplified the implementation by simply always initializing
the mutex. For a webserver, the cost overhead of initializing a mutex is
negligible. The mutex is only ever locked by Mongoose itself during
connection close, which will add no cost unless the application actually
was using the mutex--in which case, it needs that mutex to be locked.