Better komma detection in parse_header
The komma is not the most reliable way to check if a HTTP headers variable ends. Take this case:

http://x.x.x.x/?name=foo,bar

However, the current header parsing, thinks the uri variable ends after the word foo because of the komma. To properly determine if a variable ends, you should not just look at the komma, but instead check the komma space sequence. Spaces are not allowed in HTTP headers so they will automatically be translated to %20. Therefor, you can be sure that a komma space sequence actually means the end of the variable.

This bug made it impossible to authenticate when using these kinds of GET urls.