Add optional flag to request immediate sending of data
Add a user-settable flag MG_F_SEND_IMMEDIATELY that requests mg_send()
to immediately send the provided data if the write to the socket would
not block.  Otherwise the data is just appended to the send buffer but
is not sent.

This flag is not maintained in the flags variable so it must be set
before calling mg_send() each time if immediate sending is desired.

To implement this change in mg_send() required moving some definitions
earlier in the file.