Allow controlling behavior when MG_MAX_RECV_SIZE is reached
When MG_MAX_RECV_SIZE is reached we may want to stop reading
rather than throwing an error as a full buffer may just mean we
need to wait for the consumer to clear the buffer on a future
poll iteration.

Each poll iteration will call MG_EV_POLL which can consume from
the buffer to allow reading to continue, we should not assume
that the buffer will always be consumed below the recv max after
each poll iteration as the consumer may need to apply backpressure
to the recv buffer if the consumer is not able to clear the recv
buffer faster than read_conn fills it.

Introduce a MG_EV_BUFFER_FULL event which allows the user to
choose what to do when the buffer is reached, with the existing
erroring behavior remaining the default.