Use '*' for a wild-card/default for mime type if no file extension matches.
This allows someone to specify a wild card mime type to be used if there is no exact match in the file extension list.  For example `*=application/octet-stream, png=image/png, txt=text/plain` the mime-type `application/octet-stream` will be used if there are no files with an extension of txt or png.

If some reason there are multiple wild card items in the list. The first one will be used.